reverse linked list

+++ 0x13-more_singly_linked_lists/102-free_listint_safe.c
@@ -6,7 +6,7 @@
  *
  * Return: number of elements in the freed list
  */
-size_t free_listint_safe(listint_t **h)
+int free_listint_safe(listint_t **h)
 {
 size_t len = 0;
 int diff;
